Rename and rellocate eslint files and update references
