+++ package.json
@@ -99,7 +99,7 @@
     "prettier-ci": "prettier --check 'src/**/*.{js,jsx}'",
     "lint": "eslint --ext .js,.jsx src",
     "lint:production": "NODE_ENV=production yarn lint",
-    "test:lint-rules": "jest tests/",
+    "test:lint-rules": "jest eslint-plugin-ato/tests",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "happo": "happo",
@@ -136,7 +136,7 @@
     "enzyme": "^3.11.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-ato": "file:./eslint",
+    "eslint-plugin-ato": "file:./eslint-plugin-ato",
     "eslint-plugin-no-only-tests": "^2.4.0",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-security": "^1.4.0",
